requirements.txt
streamlit
pandas
numpy
plotly
python-dotenv
requests
feedparser
holidays
google-genai
scikit-learn
joblib
